refactor: 닉네임 생성 관련 로직 패키지를 domain으로 이동
