+++ backend/turip-app/src/main/java/turip/account/domain/nickname/RandomNicknameCreator.java
@@ -1,9 +1,8 @@
-package turip.account.service;
+package turip.account.domain.nickname;
 
 import java.security.SecureRandom;
 import java.util.Random;
 import org.springframework.stereotype.Component;
-import turip.account.domain.NicknameWord;
 
 @Component
 public class RandomNicknameCreator implements NicknameCreator {
